(sim) Fix kslli32

+++ sim/riscv/sim-main.c
@@ -2726,7 +2726,7 @@ execute_p (SIM_CPU *cpu, unsigned_word iw, const struct riscv_opcode *op, int ex
       {
 	if (imm5u != 0)
 	  {
-	    int32_t res;
+	    int64_t res;
 	    for (i = 0; i < vec32_num; i++)
 	      {
 		res = (int64_t) *(ptr_a32 + i) << imm5u;
